src/backend/microservices/dataset_loading/requirements.txt
# Pendiente de añadir las versiones

pandas # Pandas para crear los dataframes a partir de los datos de la API
fastapi
uvicorn
python-multipart
httpx
openpyxl
redis
pickle
pyarrow
uuid
mysql.connector
python-dotenv
logging